SimpleChatBotWithMemory/requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
langchain-text-splitters
openai
python-dotenv
fastapi
langserve
sse_starlette